+++ src/avt/Expressions/ImageProcessing/avtMeanFilterExpression.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2015, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2017, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -134,7 +134,7 @@ avtMeanFilterExpression::ProcessArguments(ArgsExpr *args,
 {
     // Get the argument list and # of arguments
     std::vector<ArgExpr*> *arguments = args->GetArgs();
-    int nargs = arguments->size();
+    size_t nargs = arguments->size();
 
     // Check number of args
     if (nargs < 1 || nargs > 4)
