+++ src/avt/Expressions/TimeIterators/avtTimeIteratorDataTreeIteratorExpression.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2015, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2017, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -166,10 +166,10 @@ avtTimeIteratorDataTreeIteratorExpression::PrepareAndExecuteDataset(
 
     bool haveZonal = false;
 
-    size_t nvars = varnames.size();
+    int nvars = (int)varnames.size();
     if (cmfeType == POS_CMFE)
         nvars--;
-    for (size_t i = 0 ; i < nvars ; i++)
+    for (int i = 0 ; i < nvars ; i++)
     {
         std::string vname = GetInternalVarname(i);
         vtkDataArray *cell_data1 = ds->GetCellData()->GetArray(vname.c_str());
@@ -187,7 +187,7 @@ avtTimeIteratorDataTreeIteratorExpression::PrepareAndExecuteDataset(
     if (haveZonal)
         doZonal = true;  // mixed centering -> zonal
 
-    for (size_t i = 0 ; i < nvars ; i++)
+    for (int i = 0 ; i < nvars ; i++)
     {
         std::string vname = GetInternalVarname(i);
         vtkDataArray *cell_data1 = ds->GetCellData()->GetArray(vname.c_str());
