+++ src/avt/Filters/avtLineToPolylineFilter.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2015, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2017, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -279,7 +279,7 @@ avtLineToPolylineFilter::ExecuteData(avtDataRepresentation *inDR)
             if(shape.size() > (size_t)ptsBufSize)
             {
                 delete [] pts;
-                ptsBufSize = shape.size();
+                ptsBufSize = (int)shape.size();
                 pts = new vtkIdType[ptsBufSize];
             }
             for(size_t i = 0; i < shape.size(); ++i)
