+++ src/avt/Filters/avtThresholdFilter.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2015, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2017, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -485,7 +485,7 @@ vtkDataSet *
 avtThresholdFilter::ThresholdToPointMesh(vtkDataSet *in_ds)
 {
     const stringVector curVariables = atts.GetListedVarNames();
-    int curVarCount = curVariables.size();
+    int curVarCount = (int)curVariables.size();
     int curVarNum;
     vtkPointData *inPointData = in_ds->GetPointData();
