+++ src/avt/IVP/avtIVPNIMRODIntegrator.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2015, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2017, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -406,7 +406,7 @@ avtIVPNIMRODIntegrator::getBfield(const avtIVPField* field, avtVector y)
   }
   else 
   {
-    float B[3];
+    double B[3];
 
     m3dField->interpBcomps(B, pt, element, xieta);
