+++ src/avt/Pipeline/Pipeline/avtNamedSelection.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2015, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2017, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -1232,11 +1232,11 @@ avtLocationsNamedSelection::GetMatchingIds(vtkDataSet *ds, std::vector<vtkIdType
     vtkIdType cellId;
     
     vtkGenericCell *cell = vtkGenericCell::New();
-    int nLocs = locations.size();
+    size_t nLocs = locations.size();
     std::map<vtkIdType, int> idMap;
 
     //Use a map to ensure we don't get duplicate cells.
-    for (int i = 0; i < nLocs; i++)
+    for (size_t i = 0; i < nLocs; i++)
     {
         p[0] = locations[i].x;
         p[1] = locations[i].y;
