+++ src/databases/M3DC1/avtM3DC1Field.C
@@ -1,6 +1,6 @@
 /*****************************************************************************
 *
-* Copyright (c) 2000 - 2015, Lawrence Livermore National Security, LLC
+* Copyright (c) 2000 - 2017, Lawrence Livermore National Security, LLC
 * Produced at the Lawrence Livermore National Laboratory
 * LLNL-CODE-442911
 * All rights reserved.
@@ -242,7 +242,7 @@ int avtM3DC1Field::register_vert(std::vector< vertex > &vlist,
     // Are the two points with the tollerance?
     if (dx*dx + dy*dy < tol)
     {
-      return i;
+      return (int)i;
     }
   }
